+++ app/Http/Controllers/Dashboard/RoleController.php
@@ -18,6 +18,9 @@ class RoleController extends Controller
      */
     public function index(Request $request)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'read roles')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         $roles = Role::RolesData($request);
         return RolesResource::collection($roles);
     }
@@ -40,6 +43,9 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'write roles')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         $this->validate($request,[
             'role'=> 'required|string',
             'permissions'=> 'required'
@@ -69,6 +75,9 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'update roles')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         return $this->returnData('role',new RolesResource($role));
     }
 
@@ -81,6 +90,9 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'update roles')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         $this->validate($request,[
             'role'=> 'required|string',
             'permissions'=> 'required'
@@ -99,6 +111,9 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        if (auth('sanctum')->user()->cannot('admin', 'delete roles')) {
+            return  \response()->json(['status'=>false,'message'=>'Access Forbidden'],403);
+        }
         try{
             $role->delete();
             return $this->returnSuccessMessage('Service Deleted Succeffully');
